Add files via upload
This is all the images ive been using to test along with what ive been able to do with coltons movement class. Currently the finished rooms in the switch statement from the overworld panel function, there is no walking through walls. You can walk from the starting position which is a hallway left of the spawn zone throughout the entire left section of the dungeon. It will require you to fix the file paths for the images in order to show perfectly.